requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
transformers==4.41.2
torch>=2.1.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.1
pypdf==4.2.0
